修改价格类型

+++ database/migrations/2015_11_16_184357_create_coupons_table.php
@@ -16,7 +16,7 @@ class CreateCouponsTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->string('type');
-            $table->decimal('content', 11, 0);
+            $table->unsignedInteger('content');
             $table->string('detail');
             $table->softDeletes();
             $table->timestamps();
