+++ database/migrations/2015_11_16_193927_create_coupon_code_table.php
@@ -17,7 +17,7 @@ class CreateCouponCodeTable extends Migration
             $table->string('code');
             $table->string('name');
             $table->string('type');
-            $table->decimal('content', 11, 0);
+            $table->unsignedInteger('content');
             $table->string('detail');
             $table->softDeletes();
             $table->timestamps();
